Major cleanup... removed registration class with everything on userController now... removed redundancy
