+++ web/deleteUserPage.xhtml
@@ -17,14 +17,14 @@
         <title>Delete User Page</title>
     </h:head>
     <h:body>
-        <c:if test="#{registration.deleted}">
+        <c:if test="#{userController.deleted}">
             <h1>Unreal War</h1><br></br>
             <h3>We are sorry to see you go</h3>
             <h:form>
-                <h:commandButton value="Return to login page" class="btn btn-lg raised" style="font-size:24px" action="#{registration.clearFields()}"/>
+                <h:commandButton value="Return to login page" class="btn btn-lg raised" style="font-size:24px" action="#{userController.clearFields()}"/>
             </h:form>
         </c:if>
-        <c:if test="#{!registration.deleted}">
+        <c:if test="#{!userController.deleted}">
             <div class="container panel panel-default">
                 <div class="row">
                     <div class="forum-group col-lg-8 col-lg-offset-2">
@@ -36,9 +36,9 @@
                         <div class="form-group">
                             <h:form>
                                 <p> Delete Account</p>
-                                <p>Please enter your username: <h:inputText value="#{registration.username}"/></p>
-                                <p>Confirm your password: <h:inputSecret value="#{registration.password}"/></p>
-                                <h:commandButton value="Delete User" class="btn btn-lg raised" style="font-size:24px" action="#{registration.deleteUser(registration.username, registration.password)}"/>
+                                <p>Please enter your username: <h:inputText value="#{userController.instance.username}"/></p>
+                                <p>Confirm your password: <h:inputSecret value="#{userController.instance.passhash}"/></p>
+                                <h:commandButton value="Delete User" class="btn btn-lg raised" style="font-size:24px" action="#{userController.deleteUser(userController.instance.username, userController.instance.passhash)}"/>
                             </h:form>
                         </div>
                     </div>
